# docs/swagger.yaml
openapi: 3.0.0
info:
  title: ReactViteProKit API
  description: >
    Full-featured backend API built with Node.js, Express, MongoDB, Cloudinary.  
    Modular MVC design with Auth, User, Admin, Product, File, OTP modules.
    All responses follow a standard format with success status, data, and message fields.
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:5000/api
    description: 
  - url: https://example.com/api
    description: 
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token required for authenticated endpoints
paths:
...
  /auth/verify-otp:
    post:
      tags:
        - Auth
      summary: Verify OTP and complete registration
      description: >
        Verifies the OTP code and completes user registration.
        Only requires phone and OTP code - all other data (email, password) 
        was sent in the initial registration request.
      operationId: verifyOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phone
                - code
              properties:
                phone:
                  type: string
                  description: Phone number
                  example: "+989123456789"
                code:
                  type: string
                  description: OTP code sent to phone
                  example: "123456"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        example: "507f1f77bcf86cd799439011"
                  message:
                    type: string
                    example: "Registered successfully"
        "400":
          description: Invalid or expired OTP, too many attempts, or user already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Error message"
                  message:
                    type: string
                    example: "Error details"
        "429":
          description: Too many incorrect OTP attempts
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Error message"
                  message:
                    type: string
                    example: "Error details"
...
  /auth/forgot-password:
    post:
      tags:
        - Auth
      summary: Request OTP for password reset
      description: >
        Sends an OTP code to the user's phone number for password reset.
        Rate limited to 3 requests per 5 minutes.
        OTP expires in 2 minutes.
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phone
              properties:
                phone:
                  type: string
                  description: User phone number
                  example: "+989123456789"
      responses:
        "200":
          description: OTP sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      phone:
                        type: string
                        example: "+989123456789"
                  message:
                    type: string
                    example: "OTP sent successfully"
        "400":
          description: Phone required or user not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Error message"
                  message:
                    type: string
                    example: "Error details"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Error message"
                  message:
                    type: string
                    example: "Error details"
        "429":
          description: Too many OTP requests, rate limited
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Error message"
                  message:
                    type: string
                    example: "Error details"
  /auth/reset-password:
    post:
      tags:
        - Auth
      summary: Reset password with OTP verification
      description: >
        Resets user password after verifying the OTP code.
        Requires phone, OTP code, and new password.
        Maximum 5 incorrect OTP attempts.
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phone
                - code
                - newPassword
              properties:
                phone:
                  type: string
                  description: User phone number
                  example: "+989123456789"
                code:
                  type: string
                  description: OTP code sent to phone
                  example: "123456"
                newPassword:
                  type: string
                  format: password
                  description: New password
                  example: "NewPassword123!"
      responses:
        "200":
          description: Password updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: "null"
                  message:
                    type: string
                    example: "Password updated successfully"
        "400":
          description: Invalid OTP, expired OTP, or too many attempts
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Error message"
                  message:
                    type: string
                    example: "Error details"
        "404":
          description: User or OTP not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Error message"
                  message:
                    type: string
                    example: "Error details"
        "429":
          description: Too many incorrect OTP attempts
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Error message"
                  message:
                    type: string
                    example: "Error details"
...